accessible/tests/mochitest/events/test_scroll.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>

<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />
  <script type="application/javascript"
          src="chrome://mochikit/content/chrome-harness.js"/>

  <script type="application/javascript"
          src="../common.js" />
  <script type="application/javascript"
          src="../role.js" />
  <script type="application/javascript"
          src="../states.js" />
  <script type="application/javascript"
          src="../events.js" />
  <script type="application/javascript"
          src="../browser.js"></script>

  <script type="application/javascript">
  <![CDATA[

    ////////////////////////////////////////////////////////////////////////////
    // Tests

    function getAnchorJumpInTabDocument(aTabIdx)
    {
      var tabDoc = aTabIdx ? tabDocumentAt(aTabIdx) : currentTabDocument();
      return tabDoc.querySelector("a[name='link1']");
    }

    function loadTab(aURL)
    {
      this.eventSeq = [
        new asyncInvokerChecker(EVENT_DOCUMENT_LOAD_COMPLETE, currentTabDocument),
        new asyncInvokerChecker(EVENT_SCROLLING_START, getAnchorJumpInTabDocument)
      ];

      this.invoke = function loadTab_invoke()
      {
        tabBrowser().loadURI(aURL);
      }

      this.getID = function loadTab_getID()
      {
        return "load tab: " + aURL;
      }
    }

    function loadTabInBackground(aURL)
    {
      this.eventSeq = [
        new invokerChecker(EVENT_DOCUMENT_LOAD_COMPLETE, tabDocumentAt, 1)
      ];

      this.unexpectedEventSeq = [
        new invokerChecker(EVENT_SCROLLING_START, getAnchorJumpInTabDocument, 1)
      ];

      this.invoke = function loadTabInBackground_invoke()
      {
        tabBrowser().loadOneTab(aURL, null, "", null, true);
      }

      this.getID = function loadTabInBackground_getID()
      {
        return "load tab in background: " + aURL;
      }
    }

    function switchToBackgroundTab()
    {
      this.eventSeq = [
        new invokerChecker(EVENT_SCROLLING_START, getAnchorJumpInTabDocument)
      ];

      this.invoke = function switchToBackgroundTab_invoke()
      {
        tabBrowser().selectTabAtIndex(1);
      }

      this.getID = function switchToBackgroundTab_getID()
      {
        return "switch to background tab";
      }
    }

    //gA11yEventDumpToConsole = true; // debug stuff

    var gQueue = null;
    function doTest()
    {
      var rootDir = getRootDirectory(window.location.href);

      /*
       * When tests are packed in a .jar, we need to extract them so we 
       * can access the specific url with a file:// protocol which appears
       * to be required by loadURI() (at least a file without an embedded .jar)
       */
      var jar = getJar(rootDir);
      if (jar) {
        var tmpdir = extractJarToTmp(jar);
        rootDir = "file://" + tmpdir.path + '/';
      }
      var url = rootDir + "scroll.html#link1";

      gQueue = new eventQueue();

      gQueue.push(new loadTab(url));
      gQueue.push(new loadTabInBackground(url));
      gQueue.push(new switchToBackgroundTab());
      gQueue.onFinish = function() { closeBrowserWindow(); }

      gQueue.invoke(); // Will call SimpleTest.finish();
    }

    if (!MAC) {
      SimpleTest.waitForExplicitFinish();
      openBrowserWindow(doTest);
    } else {
      todo(false, "Re-enable on Mac after fixing bug 835338");
    }
  ]]>
  </script>

  <vbox flex="1" style="overflow: auto;">
    <body xmlns="http://www.w3.org/1999/xhtml">
      <a target="_blank"
         href="https://bugzilla.mozilla.org/show_bug.cgi?id=691734"
         title="Make sure scrolling start event is fired when document receive focus">
        Mozilla Bug 691734
      </a>

      <p id="display"></p>
      <div id="content" style="display: none">
      </div>
      <pre id="test">
      </pre>
    </body>

    <vbox id="eventdump"></vbox>
  </vbox>
</window>
